+++ packages/svelte-undp-components/src/lib/components/pg_tileserv/VectorSimulation.stories.svelte
@@ -72,5 +72,5 @@
 		layerId: layer.id,
 		datasetUrl: sourceUrl
 	}}
-	children={template}
+	{template}
 />