+++ packages/svelte-undp-design/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @undp-data/svelte-undp-design
 
+## 3.0.5
+
+### Patch Changes
+
+- 1ed3137: fix: remove $bindable from toggled property of Switch component in UNDP Design System because of some issue of reactivity in svelte 5. Use onchange event instead.
+
 ## 3.0.4
 
 ### Patch Changes